# internal/swagger/docs/swagger.yaml
basePath: /
definitions:
  models.Answer:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
  models.AnswerVariant:
    properties:
      code:
        type: string
      isRight:
        type: boolean
      name:
        type: string
    type: object
  models.CheckResult:
    properties:
      isCorrect:
        type: boolean
      questionCode:
        type: string
      rightAnswerCode:
        items:
          type: string
        type: array
      userAnswerCode:
        items:
          type: string
        type: array
    type: object
  models.NoiseSource:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
  models.Quest:
    properties:
      code:
        type: string
      imagePNG:
        type: string
      noise:
        items:
          $ref: '#/definitions/models.NoiseSource'
        type: array
    type: object
  models.Question:
    properties:
      answers:
        items:
          $ref: '#/definitions/models.AnswerVariant'
        type: array
      code:
        type: string
      name:
        type: string
    type: object
  models.ResponseQuestion:
    properties:
      answers:
        items:
          $ref: '#/definitions/models.Answer'
        type: array
      code:
        type: string
      name:
        type: string
    type: object
  models.StatusThema:
    enum:
    - open
    - current
    - completed
    type: string
    x-enum-varnames:
    - StatusThemaOpen
    - StatusThemaCurrent
    - StatusThemaCompleted
  models.Thema:
    properties:
      code:
        type: string
      name:
        type: string
      status:
        $ref: '#/definitions/models.StatusThema'
    type: object
  models.UserAnswer:
    properties:
      answerCodes:
        items:
          type: string
        type: array
      question_code:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a Go API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: My Go API
  version: "1.0"
paths:
  /answers:
    get:
      consumes:
      - application/json
      description: возвращает список вопросов с вариантами ответов
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.ResponseQuestion'
            type: array
      summary: возвращает список вопросов с вариантами ответов
      tags:
      - тесты
  /check:
    post:
      consumes:
      - application/json
      description: Принимает список выбранных пользователем ответов (questionCode
        + answerCodes) и возвращает результат проверки
      parameters:
      - description: Ответы пользователя
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/models.UserAnswer'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CheckResult'
            type: array
        "400":
          description: Ошибка запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Проверка ответов пользователя
      tags:
      - тесты
  /quest:
    get:
      description: Возвращает случайно выбранный вопрос из базы
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Quest'
      summary: Случайный вопрос
      tags:
      - тесты
  /thema/{code}:
    get:
      description: Возвращает текстовый блок (InfoBlock) по теме
      parameters:
      - description: Код темы
        in: path
        name: code
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Текст темы
          schema:
            type: string
        "404":
          description: Тема не найдена
          schema:
            type: string
      summary: Полная информация о теме
      tags:
      - темы
  /thema/{code}/questions:
    get:
      description: Возвращает список вопросов для указанной темы без вариантов ответов
      parameters:
      - description: Код темы
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Question'
            type: array
        "404":
          description: Вопросы для темы не найдены
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Список вопросов по теме
      tags:
      - тесты
  /themas:
    get:
      description: Возвращает краткую информацию по всем темам курса
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Thema'
            type: array
      summary: Список тем
      tags:
      - темы
schemes:
- http
swagger: "2.0"